# services/game/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GeoGuessr School Edition – game service",
        "title": "GSSR Game API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/leaderboard": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard"
                ],
                "summary": "Leaderboard (ELO or XP)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "elo or xp (default: elo)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
...
        "/maps/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "maps"
                ],
                "summary": "Get map with floors and panorama counts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Map UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/maps.mapDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/profile": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user's public profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "maps.floorItem": {
            "type": "object",
            "properties": {
                "floor_number": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "pano_count": {
                    "type": "integer"
                }
            }
        },
        "maps.mapDetail": {
            "type": "object",
            "properties": {
                "coord_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "floors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maps.floorItem"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "x_max": {
                    "type": "number"
                },
                "x_min": {
                    "type": "number"
                },
                "y_max": {
                    "type": "number"
                },
                "y_min": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}